# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Rezervo API", "version": "0.1.0", "description": "Tenant-isolated booking API. Protected business calls require the authenticated cookie and x-business-id header." },
  "servers": [{ "url": "http://localhost:4000/api" }],
  "paths": {
...
    "/business/services": { "get": { "summary": "List tenant services", "parameters": [{ "$ref": "#/components/parameters/Tenant" }], "responses": { "200": { "description": "Services" } } }, "post": { "summary": "Create tenant service", "parameters": [{ "$ref": "#/components/parameters/Tenant" }], "responses": { "201": { "description": "Created" } } } },
    "/business/bookings": { "get": { "summary": "List tenant bookings", "parameters": [{ "$ref": "#/components/parameters/Tenant" }], "responses": { "200": { "description": "Bookings" } } } },
...
  },
  "components": {
    "parameters": {
...
      "Tenant": { "name": "x-business-id", "in": "header", "required": true, "schema": { "type": "string" }, "description": "Tenant selection, verified against the authenticated session." }
    },
...
  }
}